cbtproject/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-mysql==3.8.1
idna==2.10
joblib==0.16.0
numpy==1.19.2
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.10
gunicorn==20.0.4
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2==2.8.6 
whitenoise==5.2.0
djangorestframework==3.11.1